update rcon version (#14)
* commands: move directory

* update rcon version

* commands: using palworldrcon.Connect

* commands: startServer command using config

* watch: add description in dashboard

* i18n: add non english nickname notice

+++ src/utils/rcon.go
@@ -10,7 +10,7 @@ import (
 func RconClient() (palworldrcon.Client, error) {
 	client := palworldrcon.Client{
 		Host:          config.PALWORLD_RCON_HOST,
-		Port:          uint(config.PALWORLD_RCON_PORT),
+		Port:          config.PALWORLD_RCON_PORT,
 		AdminPassword: config.PALWORLD_ADMIN_PASSWORD,
 	}
